fix: postgres sweeping doesn't need to update job state
the sweeper atomically changes the job table state to be active,
when a job gets processed, at that point it will be updated to active.
it won't get swept if there's an advisory lock. if the lock is lost and
the job never got picked up, then it will be considered stuck and swept.